+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- UI for new cost center `businessDocument` field
+
 ## [1.2.2] - 2022-03-28
 
 ### Fixed